requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
black==22.10.0
cachetools==4.2.4
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
decorator==4.4.2
dj-database-url==1.1.0
Django==4.1.2
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-el-pagination==3.3.0
django-endless-pagination==2.0
django-etc==1.4.0
django-heroku==0.3.1
django-hitcount==1.3.5
django-mathfilters==1.0.0
django-resized==1.0.2
django-simple-pagination==1.3
django-taggit==3.0.0
django-tinymce==3.5.0
django-unicorn==0.48.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
joblib==1.2.0
lxml==4.9.1
mypy-extensions==0.4.3
mysqlclient==2.1.1
nltk==3.7
openpyxl==3.0.10
orjson==3.8.0
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.5
pytz==2022.4
regex==2022.10.31
requests==2.28.1
shortuuid==1.0.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.64.1
tzdata==2022.4
urllib3==1.26.12
whitenoise==6.2.0
